New file.
This file is a computational file that allows you to calculate the analog of the
gravitational field of several bodies with specified parameters (mass, coordinate).
The received data is recorded in a two-dimensional array and stored in the Data.txt
file for further visualization with the help of a graphical shell written in Python.

File consist of function:
	1. Main
	2. Maps_with_forces Simple_Filing
	3. Maps_with_forces Filing
	4. Print_results
	5. Print_results in file